# server/openapi.yaml
openapi: 3.1.0
info:
  title: Aria API
  version: 2.0.0
  description: |
    Aria v2 music-server API. Paths and JSON shapes are byte-compatible with the
    legacy Node server; v2 adds GET /healthz, GET /api/events (SSE), and optional
    limit/offset pagination on GET /api/tracks.

    ID scheme: track `id` = sha1 hex of the file path relative to MUSIC_DIR;
    `albumId` = sha1 hex of lowercase(album directory + "\0" + album title), the
    directory being the file's folder with a CD1/Disc 2 subfolder folded into its
    parent, and the album artist replacing the (empty) directory at library root.

    Error convention: validation failures are 400 with `{"error": "<reason>"}`;
    not-found is a bare 404 (text/plain body "Not Found", no JSON);
    upstream MusicBrainz failures are 502 with `{"error": "..."}`.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/playlists:
    get:
      tags: [playlists]
      summary: All playlists
      operationId: listPlaylists
      parameters:
        - name: profileId
          in: query
          required: false
          schema: { type: string }
          description: Filter to one profile's playlists; omitted returns everyone's.
      responses:
        '200':
          description: Playlists.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Playlist' }
    post:
      tags: [playlists]
      summary: Create a playlist
      description: Manual playlists start with empty `trackIds`; smart playlists require valid `rules`.
      operationId: createPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profileId, name]
              properties:
                profileId: { type: string }
                name: { type: string, maxLength: 200 }
                type: { type: string, enum: [manual, smart], default: manual }
                rules: { $ref: '#/components/schemas/SmartRules' }
      responses:
        '200':
          description: Created playlist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/playlists/{id}:
    patch:
      tags: [playlists]
      summary: Rename a playlist / update smart rules
      description: '`rules` is only valid on smart playlists. Bumps `updatedAt` even for empty patches.'
      operationId: patchPlaylist
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 200 }
                rules: { $ref: '#/components/schemas/SmartRules' }
      responses:
        '200':
          description: Updated playlist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [playlists]
      summary: Delete a playlist
      operationId: deletePlaylist
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/playlists/{id}/tracks:
    get:
      tags: [playlists]
      summary: Resolve a playlist to full track objects
      description: |
        Manual: tracks in stored order, ids no longer in the library silently
        skipped, duplicates preserved. Smart: rules evaluated on demand against the
        current library + plays; result sorted by artist/album/discNo/trackNo and
        capped at 500. Empty rule list yields [].
      operationId: getPlaylistTracks
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Tracks.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Track' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [playlists]
      summary: Append a track to a manual playlist
      description: Duplicates are allowed. 400 on smart playlists or unknown tracks.
      operationId: addPlaylistTrack
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trackId]
              properties:
                trackId: { $ref: '#/components/schemas/Sha1' }
      responses:
        '200':
          description: Updated playlist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/playlists/{id}/tracks/{trackId}:
    delete:
      tags: [playlists]
      summary: Remove a track from a manual playlist
      description: Removes ALL occurrences of the track id. 400 on smart playlists.
      operationId: removePlaylistTrack
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: trackId, in: path, required: true, schema: { $ref: '#/components/schemas/Sha1' } }
      responses:
        '200':
          description: Updated playlist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---- radio ------------------------------------------------------------------------
...
components:
  responses:
    Ok:
      description: Success.
      content:
        application/json:
          schema:
            type: object
            required: [ok]
            properties:
              ok: { type: boolean, const: true }
    BadRequest:
      description: Validation failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found. Plain-text body ("Not Found"), no JSON.
      content:
        text/plain:
          schema: { type: string }
...
  schemas:
    Sha1:
      type: string
      pattern: '^[0-9a-f]{40}$'
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Track:
      type: object
      description: |
        A track as served by /api/tracks — file tags merged with enrichment
        credits, edits overlaid, `path` stripped, plus derived releaseType,
        canonical genres, and user tag names. Fields absent in the file tags
        are null, not omitted (except performers/orchestra, which only appear
        after enrichment).
      required:
        - id
        - addedAt
        - title
        - artist
        - albumArtist
        - album
        - albumId
        - format
        - lossless
        - hasArt
        - releaseType
        - genres
        - tags
      properties:
        id:
          $ref: '#/components/schemas/Sha1'
          description: sha1 hex of the file path relative to MUSIC_DIR.
        addedAt:
          type: string
          format: date-time
          description: First-seen scan timestamp; carried forward across rescans.
        title:
          type: string
          description: Falls back to the relative file path when the file has no title tag.
        artist: { type: string, description: 'Defaults to "Unknown Artist".' }
        albumArtist: { type: string, description: Defaults to artist. }
        album: { type: string, description: 'Defaults to "Unknown Album".' }
        albumId:
          $ref: '#/components/schemas/Sha1'
          description: sha1 hex of lowercase(albumArtist + "\0" + album).
        composer: { type: [string, 'null'] }
        conductor: { type: [string, 'null'] }
        orchestra:
          type: string
          description: Enrichment-only; absent until a MusicBrainz credit pass finds one.
        performers:
          type: array
          description: Enrichment-only per-performer credits; absent until enriched.
          items: { $ref: '#/components/schemas/Performer' }
        work: { type: [string, 'null'], description: 'Classical work, from WORK/GROUPING tags or a "Work: Movement" title split.' }
        movement: { type: [string, 'null'] }
        mbAlbumId: { type: [string, 'null'], description: MusicBrainz release id from file tags. }
        mbRecordingId: { type: [string, 'null'] }
        mbAlbumArtistId: { type: [string, 'null'] }
        year: { type: [integer, 'null'] }
        genre: { type: [string, 'null'], description: Raw first genre tag as found in the file. }
        genres:
          type: array
          items: { type: string }
          description: Canonical genres derived from the raw tag via the /api/genres taxonomy.
        trackNo: { type: [integer, 'null'] }
        discNo: { type: [integer, 'null'] }
        duration: { type: [number, 'null'], description: 'Seconds, fractional.' }
        format: { type: string, description: 'Container name, e.g. "FLAC", "MPEG", or uppercased extension.' }
        sampleRate: { type: [integer, 'null'], description: Hz. }
        bitsPerSample: { type: [integer, 'null'] }
        channels: { type: [integer, 'null'] }
        lossless: { type: boolean }
        hasArt: { type: boolean, description: 'True when GET /api/art succeeds for this albumId.' }
        releaseType:
          type: string
          enum: [Album, EP, Single, Live, Compilation]
          description: Derived per album (MB release-group type, VA/live/EP heuristics, track count + duration); album edits win. Never persisted.
        tags:
          type: array
          items: { type: string }
          description: User tag names matching this track directly, via its album, or via artist/albumArtist — each including the tag's ancestor chain.
    Performer:
      type: object
      required: [name, role]
      properties:
        name: { type: string }
        role: { type: string, description: 'Instrument attributes joined with ", ", or "vocals"/"performer".' }
...
    Playlist:
      type: object
      description: Manual playlists carry `trackIds`; smart playlists carry `rules`. The other field is absent, not null.
      required: [id, profileId, type, name, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        profileId: { type: string }
        type: { type: string, enum: [manual, smart] }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        trackIds:
          type: array
          items: { $ref: '#/components/schemas/Sha1' }
          description: Manual only. Ordered; duplicates allowed.
        rules:
          $ref: '#/components/schemas/SmartRules'
          description: Smart only.
    SmartRules:
      type: object
      required: [match, rules]
      properties:
        match: { type: string, enum: [all, any] }
        rules:
          type: array
          maxItems: 16
          items: { $ref: '#/components/schemas/SmartRule' }
    SmartRule:
      type: object
      description: |
        Valid field/op combinations:
        title, artist, albumArtist, album, genre, composer, format, credited:
        is, isNot, contains, anyOf, allOf (anyOf/allOf take a string array, 1-30 items).
        year, playCount: is, gt, lt (numeric value).
        lossless: is (boolean, or the strings "true"/"false").
        releaseType: is, isNot.
        addedDays: within (numeric value, days).
        tag: is, isNot, anyOf, allOf (exact tag-name match).
        genre matches hierarchically against the canonical taxonomy, with raw-substring
        fallback for pre-canonicalization rules. credited matches artist, conductor,
        orchestra and performer names.
      required: [field, op, value]
      properties:
        field:
          type: string
          enum: [title, artist, albumArtist, album, genre, composer, format, credited,
                 year, lossless, releaseType, playCount, addedDays, tag]
        op:
          type: string
          enum: [is, isNot, contains, anyOf, allOf, gt, lt, within]
        value:
          oneOf:
            - { type: string }
            - { type: number }
            - { type: boolean }
            - { type: array, minItems: 1, maxItems: 30, items: { type: string, minLength: 1 } }
...